use Arc for nested raw config type

+++ mystiko_config/src/raw/contract/pool.rs
@@ -24,7 +24,7 @@ pub struct RawPoolContractConfig {
     pub contract_type: ContractType,
 
     #[validate(range(min = 1))]
-    pub start_block: u32,
+    pub start_block: u64,
 
     #[validate(range(min = 1))]
     #[builder(default = None)]
